feat(auth): choose computer during device approval
