+++ packages/platform/src/db.ts
@@ -174,6 +174,7 @@ interface DeviceCodesTable {
   device_code: string;
   user_code: string;
   clerk_user_id: string | null;
+  runtime_slot: string | null;
   expires_at: number;
   last_polled_at: number | null;
   created_at: number;
@@ -903,11 +904,13 @@ async function migrate(db: Kysely<PlatformDatabase>): Promise<void> {
       device_code TEXT PRIMARY KEY,
       user_code TEXT NOT NULL UNIQUE,
       clerk_user_id TEXT,
+      runtime_slot TEXT,
       expires_at BIGINT NOT NULL,
       last_polled_at BIGINT,
       created_at BIGINT NOT NULL
     )
   `.execute(db);
+  await sql`ALTER TABLE device_codes ADD COLUMN IF NOT EXISTS runtime_slot TEXT`.execute(db);
   await sql`CREATE INDEX IF NOT EXISTS idx_device_codes_user_code ON device_codes(user_code)`.execute(db);
   await sql`CREATE INDEX IF NOT EXISTS idx_device_codes_expires_at ON device_codes(expires_at)`.execute(db);
