+++ www/content/docs/cli.mdx
@@ -48,7 +48,7 @@ You can also run without installing using `npx --yes @finnaai/matrix <command>` 
 matrix login
 ```
 
-This opens a browser for the device authentication flow. If you do not have a Matrix account yet, `matrix login` waits while the browser walks you through signup, checkout, and provisioning. Approve the CLI in that same browser tab once provisioning finishes.
+This opens a browser for the device authentication flow. The confirmation page shows the signed-in account and lets you choose which available Matrix computer the CLI may access. If you do not have a Matrix account yet, `matrix login` waits while the browser walks you through signup, checkout, and provisioning. Approve the CLI in that same browser tab once provisioning finishes.
 
 **Flags**
